+++ extensions/figma-files/src/oauth.ts
@@ -19,6 +19,6 @@ export const figma = new OAuthService({
     "https://oauth.raycast.com/v1/token/rsuQtsavqnjuxmA3vCYSGa3E-D8_ruhzY4OqJIOA7aB4_wGaFh2WE0mQc_uMZ7E5Nk8-pAeW3JXE3WlUfgqu_-zNiM4yWU7i_z2-u0pO8HJFh7_H6ohVBh8fcDpk64nT",
   refreshTokenUrl:
     "https://oauth.raycast.com/v1/refresh-token/qJfyXOqhjDouaj06_54vl2O3NoIfY36R_-OOExXZNAS073Bih0aeNaHLO9xEpW6lbooqWCpT6zO7zLvbTx1MtXF2dU5d4B_of5d05Yxh27JIAPHG0uBw7fINhej_ViQ-sbE",
-  scope: "files:read",
+  scope: "file_content:read,file_metadata:read,file_versions:read,projects:read",
   personalAccessToken: PERSONAL_ACCESS_TOKEN,
 });