+++ extensions/ivpn/src/utils/objects.ts
@@ -1,0 +1,26 @@
+export function isObject(item: unknown): item is Record<string, unknown> {
+  return item !== null && typeof item === "object" && !Array.isArray(item);
+}
+
+/**
+ * Deep merge two or more objects.
+ * @param target
+ * @param sources
+ */
+export function mergeDeep<T extends Record<string, any>>(target: T, ...sources: any[]): T {
+  if (!sources.length) return target;
+  const source = sources.shift();
+
+  if (isObject(target) && isObject(source)) {
+    for (const key in source) {
+      if (isObject(source[key])) {
+        if (!target[key]) Object.assign(target, { [key]: {} });
+        mergeDeep(target[key], source[key]);
+      } else {
+        Object.assign(target, { [key]: source[key] });
+      }
+    }
+  }
+
+  return mergeDeep(target, ...sources);
+}